+++ src/components/StoreDetail/StoreDetail.jsx
@@ -6,6 +6,8 @@ import Modal from "react-modal";
 import StoreForm from "../StoreForm/StoreForm";
 import ProductStores from "../ProductStores/ProductStores";
 import { toast } from "react-toastify";
+import { confirmAlert } from 'react-confirm-alert';
+import 'react-confirm-alert/src/react-confirm-alert.css';
 
 const StoreDetail = () => {
     const { id: storeId } = useParams();
@@ -16,15 +18,14 @@ const StoreDetail = () => {
     const products = useSelector((state) => state.products.data);
     const storesForProduct = useSelector((state) => state.storesForProduct);
     const [hasProduct, setHasProduct] = useState(false);
+    const [isModalOpen, setIsModalOpen] = useState(false);
 
     useEffect(() => {
         const storeHasProduct = products?.some((product) => {
             return storesForProduct[product.id]?.some(store => store.id === storeId)
-        })
+        });
         setHasProduct(storeHasProduct);
     }, [products, storesForProduct, storeId]);
-
-    const [isModalOpen, setIsModalOpen] = useState(false);
 
     useEffect(() => {
         dispatch(getStoreDetail(storeId));
@@ -35,17 +36,44 @@ const StoreDetail = () => {
     };
 
     const closeModal = () => {
-        dispatch(getStoreDetail(storeId))
+        dispatch(getStoreDetail(storeId));
         setIsModalOpen(false);
     };
 
     const handleDeleteStore = (storeId) => {
-        if (!hasProduct) {
-            dispatch(deleteStore(storeId));
-            navigate('/stores');
-        } else {
-            toast.error('Cannot delete store with associated products');
-        }
+        confirmAlert({
+            customUI: ({ onClose }) => {
+                return (
+                    <div className="bg-white rounded-lg shadow-md p-6 text-center">
+                        <h2 className="text-lg font-semibold mb-4">Confirm deletion</h2>
+                        <p className="text-gray-700 mb-4">Are you sure you want to delete this store?</p>
+                        <div className="flex justify-center space-x-4">
+                            <button
+                                className="bg-red-500 hover:bg-red-600 text-white py-2 px-4 rounded-md focus:outline-none focus:ring-2 focus:ring-red-400"
+                                onClick={() => {
+                                    if (!hasProduct) {
+                                        dispatch(deleteStore(storeId));
+                                        navigate('/stores');
+                                        onClose();
+                                    } else {
+                                        toast.error('Cannot delete store with associated products');
+                                        onClose();
+                                    }
+                                }}
+                            >
+                                Ok
+                            </button>
+                            <button
+                                className="bg-gray-300 hover:bg-gray-400 text-gray-800 py-2 px-4 rounded-md focus:outline-none focus:ring-2 focus:ring-gray-400"
+                                onClick={onClose}
+                            >
+                                Cancel
+                            </button>
+                        </div>
+                    </div>
+                );
+            }
+        });
     };
 
     return (
@@ -101,6 +129,6 @@ const StoreDetail = () => {
             </div>
         </div>
     );
-}
+};
 
 export default StoreDetail;
